Move mission-integration tests from assets/ to mission/
Three tests that required Mission/MissionAsset fixtures (test_asset_details_in_mission,
test_asset_commands, test_asset_command_set_form_error) move to mission/test_assets.py.
assets/tests.py now only imports from assets.models and organization.models.
